remove file functions from ESM
Administration is performed by governance, therefore it should be
exposed in the ESMom only.

The `file` functions only change values for the new ESM - the current
one remains unaffected.

Governance has the choice of bundling together calls to `free`/`burn`
and `file` to immediately redeploy an ESM with the new values.